requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Django==2.2.3
django-bootstrap4==0.0.8
django-cors-headers==3.0.2
django-tenants==2.2.3
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
kombu==4.6.3
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
vine==1.3.0
django-webpack-loader==0.6.0
